refactor: extract Tier 1 shared input types into support/
Follow the Option extraction with the mechanical Tier-1 shared-type moves, all
into support/ (single source, no component-level re-exports):

- Color / Size: SInputBase's scales -> support/InputBase.ts. SInputBase and every
  input that used or re-exported them now import from there; the ~17
  component-level re-exports are removed (each component's own divergent Size,
  e.g. SButton/SAvatar, is left untouched).
- Align / TextColor: the SInputText/SInputNumber (and SPagination's Align)
  duplicates -> support/InputText.ts; Table.ts's inline aligns use it too.
- FileObject / Action: the identical SInputFileUpload + item declarations ->
  support/InputFileUpload.ts (action?: Action | null, the nullable superset).
- SInputYMD/SInputHMS `Value`: aliased to the existing support/Day.ts Ymd/Hms
  instead of re-declaring the shape.

Stories/spec repointed to the support modules. No behavior change.
