+++ lib/components/SInputDropdownItem.vue
@@ -1,9 +1,7 @@
 <script setup lang="ts">
 import IconX from '~icons/ph/x'
+import { type Size } from '../support/InputBase'
 import SAvatar from './SAvatar.vue'
-import { type Size } from './SInputBase.vue'
-
-export type { Size }
 
 export type Item = ItemText | ItemAvatar
