+++ lib/components/SInputHMS.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 import { computed, ref } from 'vue'
+import { type Hms } from '../support/Day'
 import SInputBase, { type Props as BaseProps } from './SInputBase.vue'
 
-export type { Color, Size } from './SInputBase.vue'
 export interface Props extends BaseProps {
   placeholder?: Placeholder
   noHour?: boolean
@@ -13,11 +13,7 @@ export interface Props extends BaseProps {
   modelValue?: Value
 }
 
-export interface Value {
-  hour: string | null
-  minute: string | null
-  second: string | null
-}
+export type Value = Hms
 
 export interface Placeholder {
   hour?: string
