+++ lib/components/SInputRadios.vue
@@ -8,10 +8,9 @@
 >
 import { type Component, computed } from 'vue'
 import { type Validatable } from '../composables/Validation'
-import SInputBase, { type Color, type Size } from './SInputBase.vue'
+import { type Color, type Size } from '../support/InputBase'
+import SInputBase from './SInputBase.vue'
 import SInputRadio from './SInputRadio.vue'
-
-export type { Color, Size }
 
 export interface Option<ValueType extends string | number | boolean = string | number | boolean> {
   label: string
