+++ lib/support/InputBase.ts
@@ -1,0 +1,5 @@
+// The size and color scales shared by SInputBase and every input built on it.
+
+export type Size = 'sm' | 'md' | 'mini' | 'small' | 'medium'
+
+export type Color = 'neutral' | 'mute' | 'info' | 'success' | 'warning' | 'danger'
