+++ lib/support/InputText.ts
@@ -1,0 +1,6 @@
+// The text alignment and color scales shared by SInputText / SInputNumber (and
+// anything that mirrors them, e.g. SPagination's align).
+
+export type Align = 'left' | 'center' | 'right'
+
+export type TextColor = 'neutral' | 'info' | 'success' | 'warning' | 'danger'
