+++ stories/components/SInputDropdown.01_Playground.story.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
-import SInputDropdown, { type Size } from 'sefirot/components/SInputDropdown.vue'
+import SInputDropdown from 'sefirot/components/SInputDropdown.vue'
+import { type Size } from 'sefirot/support/InputBase'
 import { type OptionText } from 'sefirot/support/Option'
 import { ref } from 'vue'
