+++ stories/components/SInputDropdown.02_Avatar_Options.story.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
-import SInputDropdown, { type Size } from 'sefirot/components/SInputDropdown.vue'
+import SInputDropdown from 'sefirot/components/SInputDropdown.vue'
+import { type Size } from 'sefirot/support/InputBase'
 import { type OptionAvatar } from 'sefirot/support/Option'
 import { ref } from 'vue'
