+++ stories/components/SInputText.01_Playground.story.vue
@@ -2,10 +2,12 @@
 import IconCheckCircle from '~icons/ph/check-circle-bold'
 import IconXCircle from '~icons/ph/x-circle-bold'
 import SButton from 'sefirot/components/SButton.vue'
-import SInputText, { type Align, type Size, type TextColor } from 'sefirot/components/SInputText.vue'
+import SInputText from 'sefirot/components/SInputText.vue'
 import SSpinner from 'sefirot/components/SSpinner.vue'
 import { useData } from 'sefirot/composables/Data'
 import { useValidation } from 'sefirot/composables/Validation'
+import { type Size } from 'sefirot/support/InputBase'
+import { type Align, type TextColor } from 'sefirot/support/InputText'
 import { required } from 'sefirot/validation/rules'
 import { type DefineComponent, computed, ref } from 'vue'
